+++ src/ships/generateShips.js
@@ -1,0 +1,27 @@
+const battleships = {
+  size: 4,
+  quantity: 1,
+  name: 'battleship'
+};
+
+const cruisers = {
+  size: 3,
+  quantity: 2,
+  name: 'cruiser'
+};
+
+const destroyers = {
+  size: 2,
+  quantity: 3,
+  name: 'destroyer'
+};
+
+const submarines = {
+  size:1,
+  quantity: 4,
+  name: 'submarine'
+};
+
+const ships = [battleships, cruisers, destroyers, submarines];
+
+export default ships;
